util/requirements.txt
mercantile
mapbox_vector_tile
vt2geojson
pillow
requests
